Various
 * vthread32_create now used vthread32_create_raw
 * thread id returned instead of handle

+++ kernel/include/miros.h
@@ -17,7 +17,7 @@ void *const malloc(size_t const szb);
 void free(void *const buffer);
 // thread functions
 unsigned const vthread_get_all(uint32_t **const list);
-uint32_t const vthread_create(void *const(*thread)(void *const), void *const arg);
+uint64_t const vthread_create(void *const(*thread)(void *const), void *const arg);
 // file system functions
 fat32_t *const fs_mount(sd_context_t *const sd_context);
 int const fs_dir_get(fat32_entry_t *const dir_entry, char const path[]);
